Поправил всю логику проектов

+++ backend/app/Http/Controllers/Api/SpaceRoleController.php
@@ -15,15 +15,13 @@ class SpaceRoleController extends Controller
 {
     public function createSpaceRole(CreateSpaceRoleRequest $request)
     {
-        $space = Space::find($request->id);
+        $space = Space::find($request->space_id);
 
         $this->authorize('spaceAdmin', $space);
 
-        if(SpaceRole::where([['space_id', '=', $request->id], ['name', '=', $request->name]])->exists()){
+        if(SpaceRole::where([['space_id', '=', $request->space_id], ['name', '=', $request->name]])->exists()){
             return response()->json(['message' => 'Пространство уже содержит такую роль'], 422);
         }
-
-        $request->merge(['space_id' => $request->id]);
 
         $spaceRole = SpaceRole::create($request->validated());
 
@@ -54,7 +52,12 @@ class SpaceRoleController extends Controller
 
         $this->authorize('spaceAdmin', $space);
 
-        $spaceRole = SpaceRole::find($request->role_id);
+        $spaceRole = SpaceRole::where([['space_id', '=', $request->id],
+            ['id', '=', $request->role_id]])->first();
+
+        if(!$spaceRole){
+            return response()->json(['message' => 'Роль не относится к этому пространству'], 403);
+        }
 
         $spaceRole->delete();
